\name{PlasmaVolume}
\alias{PlasmaVolume}
\docType{data}
\title{
Measurements of Plasma Volume Using Two Sets of Normal Values.
}
\description{
Measurements of plasma volume expressed as a percentage of normal in 99 subjects, using two
alternative sets of normal values due to Nadler and Hurley.
}
\usage{PlasmaVolume}
\format{
  A data frame with 99 observations on the following 3 variables.
  \describe{
    \item{\code{Nadler}}{Plasma volume expressed as a percentage of normal using Nadler normal values.}
    \item{\code{Hurley}}{Plasma volume expressed as a percentage of normal using Hurley normal values.}
  }
}
\source{
Data originally supplied by C Dore, reprinted in Altman and Bland 1999.
}
\references{
Bland, J.M. and Altman, D.G., 1999. Measuring agreement in method comparison studies. Statistical methods in medical research, 8(2), pp.135-160.
}
\examples{
data(PlasmaVolume, package = "R4HCR")

}
\keyword{datasets}
